modify get_alias_edges_xn2v() in hetnode2vec_tf and test_hetnode2vec_tf.py and removing hetnode2vec.py and test_hetnode2vec.py
