+++ xn2v/__init__.py
@@ -7,9 +7,10 @@
 
 """
 from .csf_graph import CSFGraph
-from .hetnode2vec import N2vGraph
+from .hetnode2vec_tf import N2vGraph
 from .kW2V import kWord2Vec
 from .link_prediction import LinkPrediction
+from .link_prediction_with_validation import LinkPredictionWithValidation
 from .text_encoder import TextEncoder
 from .utils.tf_utils import TFUtilities
 from .text_coocurrence_encoder import TextCooccurrenceEncoder
@@ -19,6 +20,6 @@ from .word2vec import ContinuousBagOfWordsWord2Vec
 from .word2vec import SkipGramWord2Vec
 
 __all__ = [
-    "N2vGraph", "LinkPrediction", "CSFGraph", "TextEncoder", "TextCooccurrenceEncoder",
+    "N2vGraph", "LinkPrediction", "link_prediction_with_validation","CSFGraph", "TextEncoder", "TextCooccurrenceEncoder",
     "CBOWListBatcher", "kWord2Vec", "ContinuousBagOfWordsWord2Vec", "SkipGramWord2Vec", "SkipGramBatcher"
 ]
