+++ xn2v/hetnode2vec_tf.py
@@ -215,19 +215,21 @@ class N2vGraph:
         # log.info("source node:{}".format(src))
         # log.info("destination node:{}".format(dst))
         dsttype = dst[0]
-        dst2count = defaultdict(int)  # counts for going from current node ("dst") to
-        # nodes of a given type (g, p,d)
+        # counts for going from current node ("dst") to nodes of a given type (g, p,d)
+        dst2count = defaultdict(int)
         dst2prob = defaultdict(float)  # probs calculated from own2count
         total_neighbors = 0
-
-        # no need to explicitly sort, g returns a sorted list
+        # No need to explicitly sort, g returns a sorted list
         sorted_neighbors = g.neighbors(dst)
-
         for nbr in sorted_neighbors:
             nbrtype = nbr[0]
             dst2count[nbrtype] += 1
             total_neighbors += 1
         total_non_own_probability = 0.0
+        num_types = 0
+        for n, count in dst2count.items():  # count the number of types of edges from dst
+            if dst2count[n] != 0:
+                num_types += 1
 
         for n, count in dst2count.items():
             if n == dsttype:
@@ -235,24 +237,19 @@ class N2vGraph:
                 continue
             else:
                 # owntype is going to a different node type
-                dst2prob[n] = float(self.gamma) / float(count)
+                dst2prob[n] = float(self.gamma) / (float(count) * num_types)  # dividing by the num of type of edges
                 total_non_own_probability += dst2prob[n]
-
         if dst2count[dsttype] == 0:
             dst2prob[dsttype] = 0
         else:
-            dst2prob[dsttype] = \
-                (1 - total_non_own_probability) / float(dst2count[dsttype])
-
+            dst2prob[dsttype] = (1 - total_non_own_probability) / float(dst2count[dsttype])
         # Now assign the final unnormalized probs
         unnormalized_probs = np.zeros(total_neighbors)
         i = 0
-
         for dst_nbr in sorted_neighbors:
             nbrtype = dst_nbr[0]
             prob = dst2prob[nbrtype]
             edge_weight = g.weight(dst, dst_nbr)
-
             if dst_nbr == src:
                 unnormalized_probs[i] = prob * edge_weight / p
             elif g.has_edge(dst_nbr, src):
@@ -260,45 +257,41 @@ class N2vGraph:
             else:
                 unnormalized_probs[i] = prob * edge_weight / q
             i += 1
-
         norm_const = sum(unnormalized_probs)
         normalized_probs = [float(u_prob) / norm_const for u_prob in unnormalized_probs]
-
         return self.__alias_setup(normalized_probs)
 
     def __preprocess_transition_probs_xn2v(self) -> None:
-        """Preprocessing of transition probabilities for guiding the random walks.
-        This version uses gamma to calculate weighted skipping across a heterogeneous
-        network.
+        """Preprocessing of transition probabilities for guiding the random walks. This version uses gamma to
+        calculate weighted skipping across a heterogeneous network.
 
-        Assumption: The type of the node is encoded by its first character (e.g. g42 is
-        a gene).
+        Assumption: The type of the node is encoded by its first character (e.g. g42 is a gene).
 
         Returns:
             None.
         """
         starttime = time.time()
-        g = self.g
+        G = self.g
         alias_edges = {}
         alias_nodes = {}
-
-        for node in g.nodes():
+        for node in G.nodes():
+            # ASSUMPTION. The type of the node is encoded by its first character, e.g., g42 is a gene
+            # node_as_int = G.node_to_index_map[node]
             owntype = node[0]
-
-            # counts for going from current node ("own") to nodes of a given type(g,p,d)
-            own2count: Dict = defaultdict(int)
-            own2prob: Dict = defaultdict(float)  # probs calculated from own2count
+            own2count = defaultdict(int)  # counts for going from current node ("own") to nodes of a given type (g, p,d)
+            own2prob = defaultdict(float)  # probs calculated from own2count
             total_neighbors = 0
-
-            # g returns a sorted list of neighbors of node
-            sorted_neighbors = g.neighbors(node)
-
+            # G returns a sorted list of neighbors of node
+            sorted_neighbors = G.neighbors(node)
             for nbr in sorted_neighbors:
                 nbrtype = nbr[0]
                 own2count[nbrtype] += 1
                 total_neighbors += 1
-
             total_non_own_probability = 0.0
+            num_types = 0
+            for n, count in own2count.items():  # count the number of types of edges from dst
+                if own2count[n] != 0:
+                    num_types += 1
 
             for n, count in own2count.items():
                 if n == owntype:
@@ -306,43 +299,36 @@ class N2vGraph:
                     continue
                 else:
                     # owntype is going to a different node type
-                    own2prob[n] = float(self.gamma) / float(count)
+                    own2prob[n] = float(self.gamma) / float(count) * num_types
                     total_non_own_probability += own2prob[n]
-
             if own2count[owntype] == 0:
                 own2prob[owntype] = 0
             else:
-                own2prob[owntype] = \
-                    (1 - total_non_own_probability) / float(own2count[owntype])
-
-            # now assign the final unnormalized probs
+                own2prob[owntype] = (1 - total_non_own_probability) / float(own2count[owntype])
+            # Now assign the final unnormalized probs
             unnormalized_probs = np.zeros(total_neighbors)
             i = 0
-
             for nbr in sorted_neighbors:
                 nbrtype = nbr[0]
                 prob = own2prob[nbrtype]
-                unnormalized_probs[i] = prob * g.weight(node, nbr)
+                unnormalized_probs[i] = prob * G.weight(node, nbr)
                 i += 1
-
             norm_const = sum(unnormalized_probs)
             # log.info("norm_const {}".format(norm_const))
-            normalized_probs = \
-                [float(u_prob) / norm_const for u_prob in unnormalized_probs]
+            normalized_probs = [float(u_prob) / norm_const for u_prob in unnormalized_probs]
             alias_nodes[node] = self.__alias_setup(normalized_probs)
-
-        for edge in g.edges():
-            alias_edges[edge] = self.get_alias_edge_xn2v(edge[0], edge[1])
+        for edge in G.edges():
+            # src_as_int = G.node_to_index_map[edge[0]]
+            # dst_as_int = G.node_to_index_map[edge[1]]
+            # alias_edges[(src_as_int, dst_as_int)] = self.get_alias_edge_xn2v(edge[0], edge[1])
+            alias_edges[(edge[0], edge[1])] = self.get_alias_edge_xn2v(edge[0], edge[1])
 
         self.alias_edges = alias_edges
         self.alias_nodes = alias_nodes
         endtime = time.time()
         duration = endtime - starttime
-
-        log.info("Setup alias probabilities for graph in {:.2f} seconds."
-                 .format(duration))
-        print("Setup alias probabilities for graph in {:.2f} seconds."
-              .format(duration))
+        log.info("Setup alias probabilities for graph in {:.2f} seconds.".format(duration))
+        print("Setup alias probabilities for graph in {:.2f} seconds.".format(duration))
 
         return None
